# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Tezza Lending Application API
  description: "Loan products, customer limits, loan disbursement, repayments, overdue\
    \ sweeps, and notifications."
  contact:
    name: Tezza Case Study
  license:
    name: Case Study
  version: v1
servers:
- url: http://127.0.0.1:8081
  description: Generated server url
paths:
...
  /api/v1/loans:
    get:
      tags:
      - loan-controller
      summary: List loans
      operationId: list_2
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          default: id
      - name: sortDirection
        in: query
        required: false
        schema:
          type: string
          default: desc
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadPagedResponseLoanResponse"
    post:
      tags:
      - loan-controller
      summary: Disburse a loan
      operationId: disburse
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadLoanResponse"
  /api/v1/loans/{id}/repayments:
    post:
      tags:
      - loan-controller
      summary: Record a loan repayment
      operationId: repay
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadRepaymentResponse"
  /api/v1/loans/sweeps/overdue:
    post:
      tags:
      - loan-controller
      summary: Run the overdue sweep manually for a business date
      operationId: sweepOverdue
      parameters:
      - name: businessDate
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadListLoanResponse"
  /api/v1/customers:
    get:
      tags:
      - customer-controller
      summary: List customers
      operationId: list_3
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          default: id
      - name: sortDirection
        in: query
        required: false
        schema:
          type: string
          default: desc
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadPagedResponseCustomerResponse"
    post:
      tags:
      - customer-controller
      summary: Create a customer profile
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadCustomerResponse"
  /api/v1/loans/{id}/write-off:
    patch:
      tags:
      - loan-controller
      summary: Write off a loan
      operationId: writeOff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadLoanResponse"
  /api/v1/loans/{id}/cancel:
    patch:
      tags:
      - loan-controller
      summary: Cancel an open or overdue loan
      operationId: cancel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadLoanResponse"
  /api/v1/customers/{id}/loan-limit:
    patch:
      tags:
      - customer-controller
      summary: Update a customer loan limit
      operationId: updateLimit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanLimitRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadCustomerResponse"
...
  /api/v1/loans/{id}:
    get:
      tags:
      - loan-controller
      summary: Get one loan
      operationId: get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadLoanResponse"
    delete:
      tags:
      - loan-controller
      summary: Delete a loan without repayments
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadVoid"
  /api/v1/customers/{id}:
    get:
      tags:
      - customer-controller
      summary: Get one customer
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadCustomerResponse"
    delete:
      tags:
      - customer-controller
      summary: Delete a customer profile
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponsePayloadVoid"
...
components:
  schemas:
...
    Header:
      type: object
      properties:
        requestRefId:
          type: string
        responseMessage:
          type: string
        responseCode:
          type: integer
          format: int32
        customerMessage:
          type: string
        timestamp:
          type: string
...
    LoanRequest:
      required:
      - customerId
      - principalAmount
      - productId
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        principalAmount:
          minimum: 1.00
          exclusiveMinimum: false
          type: number
        structure:
          type: string
          enum:
          - LUMP_SUM
          - INSTALLMENT
        installmentCount:
          type: integer
          format: int32
        consolidatedBilling:
          type: boolean
        consolidatedDueDate:
          type: string
          format: date
    InstallmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        installmentNumber:
          type: integer
          format: int32
        dueDate:
          type: string
          format: date
        amountDue:
          type: number
        amountPaid:
          type: number
        paid:
          type: boolean
    LoanResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        loanNumber:
          type: string
        customerId:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        structure:
          type: string
          enum:
          - LUMP_SUM
          - INSTALLMENT
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
          - CANCELLED
          - OVERDUE
          - WRITTEN_OFF
        principalAmount:
          type: number
        outstandingAmount:
          type: number
        totalFeesApplied:
          type: number
        originationDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        consolidatedBilling:
          type: boolean
        consolidatedDueDate:
          type: string
          format: date
        lastFeeAccrualDate:
          type: string
          format: date
        lateFeeApplied:
          type: boolean
        installments:
          type: array
          items:
            $ref: "#/components/schemas/InstallmentResponse"
    ResponsePayloadLoanResponse:
      type: object
      properties:
        header:
          $ref: "#/components/schemas/Header"
        body:
          $ref: "#/components/schemas/LoanResponse"
    RepaymentRequest:
      required:
      - amount
      type: object
      properties:
        amount:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        channel:
          type: string
        externalReference:
          type: string
    RepaymentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        loanNumber:
          type: string
        amount:
          type: number
        appliedAmount:
          type: number
        overpaymentAmount:
          type: number
        channel:
          type: string
        externalReference:
          type: string
        paidAt:
          type: string
          format: date-time
    ResponsePayloadRepaymentResponse:
      type: object
      properties:
        header:
          $ref: "#/components/schemas/Header"
        body:
          $ref: "#/components/schemas/RepaymentResponse"
    ResponsePayloadListLoanResponse:
      type: object
      properties:
        header:
          $ref: "#/components/schemas/Header"
        body:
          type: array
          items:
            $ref: "#/components/schemas/LoanResponse"
    CustomerRequest:
      required:
      - email
      - firstName
      - lastName
      - phoneNumber
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        loanLimit:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
        riskGrade:
          type: string
        preferredNotificationChannel:
          type: string
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        loanLimit:
          type: number
        riskGrade:
          type: string
        preferredNotificationChannel:
          type: string
        registrationOtpExpiresAt:
          type: string
          format: date-time
    ResponsePayloadCustomerResponse:
      type: object
      properties:
        header:
          $ref: "#/components/schemas/Header"
        body:
          $ref: "#/components/schemas/CustomerResponse"
    LoanLimitRequest:
      required:
      - loanLimit
      type: object
      properties:
        loanLimit:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
...
    PagedResponseLoanResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LoanResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    ResponsePayloadPagedResponseLoanResponse:
      type: object
      properties:
        header:
          $ref: "#/components/schemas/Header"
        body:
          $ref: "#/components/schemas/PagedResponseLoanResponse"
    PagedResponseCustomerResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CustomerResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    ResponsePayloadPagedResponseCustomerResponse:
      type: object
      properties:
        header:
          $ref: "#/components/schemas/Header"
        body:
          $ref: "#/components/schemas/PagedResponseCustomerResponse"
    ResponsePayloadVoid:
      type: object
      properties:
        header:
          $ref: "#/components/schemas/Header"
        body:
          type: object